+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+
+### Changed
+- Renamed CreateUpdateKYCRequestBuilder -> CreateChangeRoleRequestBuilder
+
+### Removed
+- Passing keyLevelToSet to CreateChangeRoleRequestBuilder.createChangeRoleRequest
+
 ## [1.0.0-rc.1] - 2019-02-20
 ### Added
 - `AssetPairRecord` for wrapping asset pairs